+++ src/plugin.ts
@@ -1,7 +1,7 @@
 // Part of React Advanced Odontogram - https://github.com/ZoliQua/React-Odontogram-Modul
 // Created by Zoltan Dul (https://github.com/ZoliQua) 2025-2026
 
-import type { Language } from "./i18n/translations";
+import type { Language } from "./i18n/languages";
 
 /**
  * Defines the SVG layer where a plugin's visual output will be inserted.